+++ forge/src/main/java/nl/curryducker/seamless/forge/mixin/BrimstoneClusterBlockMixin.java
@@ -1,0 +1,41 @@
+package nl.curryducker.seamless.forge.mixin;
+
+import biomesoplenty.common.block.BrimstoneClusterBlock;
+import net.minecraft.core.BlockPos;
+import net.minecraft.world.item.DyeColor;
+import net.minecraft.world.level.BlockGetter;
+import net.minecraft.world.level.block.BedBlock;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.block.state.properties.DoubleBlockHalf;
+import net.minecraft.world.level.block.state.properties.EnumProperty;
+import net.minecraft.world.phys.shapes.CollisionContext;
+import net.minecraft.world.phys.shapes.VoxelShape;
+import nl.curryducker.seamless.SeamlessShapes;
+import org.spongepowered.asm.mixin.Final;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
+
+@Mixin(BrimstoneClusterBlock.class)
+public class BrimstoneClusterBlockMixin extends BedBlock {
+    @Shadow @Final public static EnumProperty<DoubleBlockHalf> HALF;
+
+    public BrimstoneClusterBlockMixin(DyeColor arg, Properties arg2) {
+        super(arg, arg2);
+    }
+
+    @Inject(method = "getShape", at = @At("HEAD"), cancellable = true)
+    public void getShape(BlockState blockState, BlockGetter blockGetter, BlockPos blockPos, CollisionContext collisionContext, CallbackInfoReturnable<VoxelShape> cir) {
+        cir.setReturnValue(SeamlessShapes.brimstoneClusterBlock(blockState.getValue(HALF) == DoubleBlockHalf.LOWER));
+    }
+
+    @Override
+    public VoxelShape getCollisionShape(BlockState blockState, BlockGetter blockGetter, BlockPos blockPos, CollisionContext collisionContext) {
+        VoxelShape SHAPE = Block.box(3.0, 0.0, 3.0, 13.0, 16.0, 13.0);
+        VoxelShape SHAPE_TOP = Block.box(6.0, 0.0, 6.0, 10.0, 8.0, 10.0);
+        return blockState.getValue(HALF) == DoubleBlockHalf.UPPER ? SHAPE_TOP : SHAPE;
+    }
+}